[#84][news] refactor: common 에러핸들러 적용
