+++ news/src/main/java/com/newpick4u/news/global/config/RedisConfig.java
@@ -1,9 +1,8 @@
-package com.newpick4u.news.news.infrastructure.redis;
+package com.newpick4u.news.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.RedisPassword;
 import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
